+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Searchlight does its best to use [semantic versioning](http://semver.org), for t
 
 ## Unreleased
 
-Nothing
+- Option readers now will find symbol keys as well as string keys
 
 ## v4.0.0 - 2015-10-28